requirements.txt
streamlit
langchain
sentence-transformers
torch
transformers
PyPDF2
python-docx
pandas
matplotlib
